vector/requirements.txt
# Aerospike dependencies
--extra-index-url https://aerospike.jfrog.io/artifactory/api/pypi/aerospike-pypi-dev/simple
aerospike-vector-search
aerospike

# Web Server 
fastapi
uvicorn[standard]
python-multipart

# Vector dependencies
sentence_transformers
torch
Pillow